requirements.txt
# Findr Backend Dependencies

# Core framework
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.10.0
python-dotenv>=1.0.0

# AI / LLM
openai>=1.50.0

# YouTube
yt-dlp>=2024.10.0
youtube-transcript-api>=0.6.0

# HTTP
requests>=2.31.0

# Database (real-time)
convex>=0.7.0

# Browser infrastructure — Daytona sandboxes for headless screenshot capture
daytona>=0.10.0

# Browser infrastructure — Browser Use Skills for TikTok/X search
browser-use-sdk>=3.0.0

# Playwright — used inside Daytona sandboxes for YouTube screenshot capture
# Not needed on the host machine, only inside the sandbox runtime.
# Listed here for documentation; the sandbox installs it at runtime.
# playwright>=1.40.0

# Editr — Google Cloud AI (Gemini, Imagen, Lyria)
# Lyria RealTime uses the live music websocket APIs present in newer SDK releases.
google-genai>=1.24.0
numpy>=1.26.0
librosa>=0.11.0
pydub>=0.25.1
Pillow>=10.4.0

# Editr — Google Cloud Storage (edited video upload + signed URLs)
google-cloud-storage>=2.14.0

# Editr — PySceneDetect (edit level detection, runs inside Daytona sandbox)
# Listed for documentation; the custom sandbox snapshot has it pre-installed.
# scenedetect[opencv]>=0.6.0
